+++ salesforce/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Salesforce Connector Change Log
 
+## 2022/7/25
+* Improved handling of permission sets to handle associations to unkown users
+
 ## 2022/4/18
 * OS environment variable `COOKIE_API_TOKEN` changed to `VEZA_API_KEY`